feat(wasm): Add test against minified browser bundle (#4526)
The current wasm integration test uses the browser bundle, and is therefore a quick smoke test of whether the bundle is legit. At least temporarily, this makes it test against the minified bundle, too, so it can act as a similar smoke test there. This is helpful as we're changing the bundling process, as it's an extra check to make sure we haven't broken the world.

Because this is a likely-temporary (ab)use of this integration test, it's done in the quickest and dirtiest way possible: the page which gets loaded is replicated, with the SDK bundle pointing to the minified version rather than the regular version. If we decide to keep testing against both bundles, we should do it in a more elegant way.